Implement "Miscellaneous" shortcuts. No MenuBar.
The menubar makes use of private interfaces (see:
nsMenuBar(Frame|Listener).cpp) and the approach
used in Toolbars.jsm is not compatible with the
regular shortcut. We would also require a way to
determine the current menubar state and
menubar.visible only relates whether the menubar
will display at all.

The button toggleReaderMode opens a given URL
in a new tab and calls our implementation after
4 seconds. A better solution may be to add a
listener to the new tab and wait for it to load
before calling toggleReaderMode().

The implementation of the caret browsing shortcut
is hardcoded in browser.xml; we've adapted it for
our purposes.